DSA3101_Projects/Docker_DSA3101/requirements.txt
flask
email_validator
mysql-connector-python
